+++ src/colmedspa.jl
@@ -1,5 +1,6 @@
-## Translation of function 'spatial.median' available in the script "PcaLocantore.R"
-## of package rrcov v.1.4-3 on R CRAN (Thanks to V. Todorov, 2016)
+## Translation of function 'spatial.median' available in the 
+## script "PcaLocantore.R" of package rrcov v.1.4-3 on R CRAN 
+## (Thanks to V. Todorov, 2016)
 function colmedspa(X; delta = 1e-6) 
     X = ensure_mat(X)
     n, p = size(X)